+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## 1.0.0
 
-- Initial version.
+# Changelog
+
+## [1.0.0] - 2025-02-05
+### Added
+- Initial release with basic functionality.
+- Create Model class, Controller class, Service class and correspondent Widget.
+
+## [1.1.0] - 2025-02-13
+### Changed
+- Updated the State, Controller and Widget structure.
+- Added Example for the new Implementation.